+++ packages/contracts/src/http/concerns/interacts-with-state.ts
@@ -1,7 +1,7 @@
 
-import { StateBag } from './state-bag.js'
+import { StateBag, HttpStateData } from './state-bag.js'
 
-export interface InteractsWithState {
+export interface InteractsWithState<State = HttpStateData> {
   /**
    * Share a given `state` across HTTP requests. Any previously
    * set state will be merged with the given `state`.
@@ -11,6 +11,8 @@ export interface InteractsWithState {
    * response.share({ user: { id: 1, name: 'Marcus' } })
    * ```
    */
+  share<K extends keyof State> (key: K, value: State[K]): this
+  share (values: Partial<State>): this
   share (key: string | any, value?: any): this
 
   /**
@@ -23,5 +25,5 @@ export interface InteractsWithState {
    * // something like "{ app: {…}, user: {…} }"
    * ```
    */
-  state (): StateBag
+  state (): StateBag<State>
 }